feat: further tweaks in restructuring
Now, the sidebar opens and closes instead of the nav within.
Users should just be able to overwrite the sidebar block while
still keeping the functionality.
