+++ src/sphinxawesome_theme/layout.html
@@ -92,7 +92,7 @@
       </header>
 
       {%- if theme_show_nav|lower == 'true' -%}
-        <aside id="sidebar" class="grid-area-sidebar print:hidden overflow-y-auto">
+        <aside id="sidebar" class="grid-area-sidebar print:hidden overflow-y-auto transition-all transform transform-gpu -translate-x-full opacity-0 duration-300 md:translate-x-0 md:opacity-100">
           {% block sidebar %}
             {% include "nav.html" %}
           {% endblock %}
@@ -103,22 +103,27 @@
         {%- if pagename != master_doc and theme_show_breadcrumbs|lower == 'true' %}
           {%- include "breadcrumbs.html" %}
         {% endif -%}
+
         {#-
             for normal pages, the "body" block is rendered in `page.html` which is
             inherited unmodified from the "basic" theme.
         -#}
+
         <article role="main" class="flex-1">
           {% block body %}{% endblock %}
         </article>
+
         {%- if theme_show_prev_next|lower == 'true' %}
           {%- include "prev_next.html" %}
         {%- endif %}
+
         <footer>
           {% block footer %}
             {% include "footer.html" %}
           {% endblock %}
         </footer>
       </main>
+
       {% include "menu.html" %}
       {% include "search-pane.html" %}
       {% include "modals.html" %}
