+++ src/sphinxawesome_theme/nav.html
@@ -5,9 +5,9 @@
 
 <nav
   role="navigation"
-  class="h-full w-sidebar fixed md:relative inset-y-0 left-0 z-20 md:z-0 bg-white text-gray-700 border-r overflow-y-auto pt-16 flex flex-col transition-transform transform transform-gpu -translate-x-full duration-300 md:translate-x-0"
+  class="h-full w-sidebar fixed md:relative inset-y-0 left-0 z-20 md:z-0 bg-white text-gray-700 border-r overflow-y-auto pt-16 flex flex-col"
 >
-  <div class="nav-toc">
+  <div class="nav-toc flex-1 px-5">
     {%- if theme_nav_include_hidden|lower == 'true' -%}
       {{ toctree(maxdepth=3, collapse=False, includehidden=true) }}
     {%- else -%}
