+++ src/sphinxawesome_theme/search-pane.html
@@ -4,8 +4,8 @@
 
 
 <div
-  id="search-pane" data-menu="closed"
-  class="w-full fixed bottom-0 left-0 z-10 bg-gray-900 overflow-x-hidden pt-10 flex flex-col items-center md:hidden"
+  id="search-pane"
+  class="w-full fixed bottom-0 left-0 z-10 bg-gray-900 overflow-x-hidden pt-10 flex flex-col items-center md:hidden transition-all opacity-0 transform transform-gpu translate-y-full duration-300"
 >
 <h4 id="search-label" class="text-xl text-white mb-4">{{ _('Enter search term') }}</h4>
