+++ tests/test_copyright.py
@@ -27,7 +27,7 @@ def test_empty_copyright_info(app: Sphinx) -> None:
     """It does not contain any Copyright info in the footer."""
     app.build()
     tree = parse_html(app.outdir / "index.html")
-    footer = tree("footer", role="contentinfo")
+    footer = tree("footer")
     assert len(footer) == 1
 
     # search for text in the children of footer
@@ -45,7 +45,7 @@ def test_dont_show_copyright(app: Sphinx) -> None:
     """It contains any Copyright info in the footer."""
     app.build()
     tree = parse_html(app.outdir / "index.html")
-    footer = tree("footer", role="contentinfo")
+    footer = tree("footer")
     assert len(footer) == 1
 
     # search for text in the children of footer
@@ -63,7 +63,7 @@ def test_show_copyright(app: Sphinx) -> None:
     """It contains any Copyright info in the footer."""
     app.build()
     tree = parse_html(app.outdir / "index.html")
-    footer = tree("footer", role="contentinfo")
+    footer = tree("footer")
     assert len(footer) == 1
 
     # search for text in the children of footer
